Data for new modeler (4.1) : moving linear problem data (from api to dataSeries) + cleaning
