+++ src/solver/modeler/CMakeLists.txt
@@ -23,15 +23,16 @@ target_link_libraries(modeler-lib
         INTERFACE
         Antares::loadModelerFiles
         Antares::modelerParameters
-	Antares::optim-model-filler
-	Antares::modeler_api
-	# TODO FIXME don't depend on implementations
-	Antares::modeler-ortools-impl
+        Antares::optim-model-filler
+        Antares::modeler_api
+        # TODO FIXME don't depend on implementations
+        Antares::modeler-ortools-impl
 )
 
 target_link_libraries(antares-modeler
         PRIVATE
         modeler-lib
+        data-series-lib
 )
 
 import_std_libs(antares-modeler)
