+++ src/solver/modeler/api/CMakeLists.txt
@@ -8,13 +8,12 @@ set(SRC_API
         include/antares/solver/modeler/api/hasBounds.h
         include/antares/solver/modeler/api/hasName.h
 
+        include/antares/solver/modeler/api/ILinearProblemData.h
         include/antares/solver/modeler/api/linearProblem.h
-        include/antares/solver/modeler/api/linearProblemData.h
         include/antares/solver/modeler/api/linearProblemFiller.h
         include/antares/solver/modeler/api/linearProblemBuilder.h
 
-        linearProblemData.cpp
-		linearProblemBuilder.cpp
+        linearProblemBuilder.cpp
 )
 
 add_library(${PROJ} ${SRC_API})
@@ -22,12 +21,7 @@ add_library(Antares::${PROJ} ALIAS ${PROJ})
 
 set_target_properties(${PROJ} PROPERTIES LINKER_LANGUAGE CXX)
 
-target_link_libraries(${PROJ}
-		PUBLIC
-		data-series-lib
-)
-
 target_include_directories(${PROJ}
         PUBLIC
-	$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
 )
