+++ src/solver/modeler/api/include/antares/solver/modeler/api/linearProblemFiller.h
@@ -24,7 +24,7 @@
 #include <memory>
 
 #include <antares/solver/modeler/api/linearProblem.h>
-#include <antares/solver/modeler/api/linearProblemData.h>
+#include "antares/solver/modeler/api/ILinearProblemData.h"
 
 namespace Antares::Solver::Modeler::Api
 {
