+++ src/solver/modeler/dataSeries/CMakeLists.txt
@@ -1,21 +1,24 @@
 
-
-set(solver_modeler_data_series "${CMAKE_SOURCE_DIR}/solver/modeler/dataSeries")
-
 set(SRC_DATA_SERIES
         include/antares/solver/modeler/dataSeries/dataSeries.h
         include/antares/solver/modeler/dataSeries/timeSeriesSet.h
-
         include/antares/solver/modeler/dataSeries/dataSeriesRepo.h
         include/antares/solver/modeler/dataSeries/scenarioGroupRepo.h
+        include/antares/solver/modeler/dataSeries/linearProblemData.h
 
         timeSeriesSet.cpp
         dataSeriesRepo.cpp
         scenarioGroupRepo.cpp
+        linearProblemData.cpp
 )
 
 add_library(data-series-lib ${SRC_DATA_SERIES})
 set_target_properties(data-series-lib PROPERTIES LINKER_LANGUAGE CXX)
+
+target_link_libraries(data-series-lib
+        PUBLIC
+        modeler_api
+)
 
 target_include_directories(data-series-lib
         PUBLIC
