+++ src/tests/src/solver/modeler/dataSeries/CMakeLists.txt
@@ -7,9 +7,6 @@ add_boost_test(unit-tests-for-modeler-data-series
         testTimeSeriesSet.cpp
         testDataSeriesRepo.cpp
         testLinearProblemData.cpp
-        INCLUDE
-        "${solver_modeler_data_series}"
         LIBS
         data-series-lib
-        Antares::modeler_api # gp : to be removed
         test_utils_unit)
