+++ src/tests/src/solver/optimisation/adequacy_patch/CMakeLists.txt
@@ -1,11 +1,7 @@
 include(${CMAKE_SOURCE_DIR}/tests/macros.cmake)
 
-# Useful variables definitions
-set(src_solver_optimisation "${CMAKE_SOURCE_DIR}/solver/optimisation")
-
 add_boost_test(tests-adq-patch
-  SRC adequacy_patch.cpp
-  INCLUDE "${src_solver_optimisation}"
-  LIBS
-  model_antares
-  array)
+        SRC adequacy_patch.cpp
+        LIBS
+        model_antares
+        array)
